requirements.txt
boto3==1.34.34
requests==2.31.0
pandas==2.1.4
gspread==5.12.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0 